refactor: isolate interactive runtime coordinator
